Applied patch for multiple file selection, documented.   Fixed some compilation warnings,   Minor change to OpenGL code.   Documented wxImageHandler::GetImageCount() and related.
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@4784 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/gtk/font.cpp
@@ -188,7 +188,7 @@ wxFont::wxFont( const wxString& fontname, const wxFontData& fontdata )
 
     tmp = tn.GetNextToken();                     // pointsize
     long num = wxStrtol (tmp.c_str(), (wxChar **) NULL, 10);
-    M_FONTDATA->m_pointSize = num / 10;
+    M_FONTDATA->m_pointSize = (int)(num / 10);
 
     tn.GetNextToken();                           // x-res
     tn.GetNextToken();                           // y-res
